# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NeptunePlay Integration API",
        "description": "Proxy endpoints for NeptunePlay integration and Seamless Wallet callbacks",
        "version": "2.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development"
        }
    ],
    "paths": {
...
        "/api/games/list": {
            "post": {
                "tags": [
                    "Games"
                ],
                "summary": "Get game list from a vendor (POST)",
                "operationId": "ce5fbaed84f8e29105327f8e47bde758",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vendorCode"
                                ],
                                "properties": {
                                    "vendorCode": {
                                        "type": "string",
                                        "example": "PGS"
                                    },
                                    "language": {
                                        "type": "string",
                                        "example": "en"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Game list retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "errorCode": {
                                            "type": "integer",
                                            "example": 0
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "503": {
                        "description": "Vendor or game under maintenance"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/game/detail": {
            "post": {
                "tags": [
                    "Games"
                ],
                "summary": "Get game detail",
                "operationId": "439a7961cb785fee5e448cfc3519fa88",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vendorCode",
                                    "gameCode"
                                ],
                                "properties": {
                                    "vendorCode": {
                                        "type": "string",
                                        "example": "PGS"
                                    },
                                    "gameCode": {
                                        "type": "string",
                                        "example": "fortune-tiger"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Game detail retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "object"
                                        },
                                        "errorCode": {
                                            "type": "integer",
                                            "example": 0
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/game/launch-url": {
            "post": {
                "tags": [
                    "Games"
                ],
                "summary": "Get game launch URL",
                "operationId": "82f0a95317eab9a9ac8767582a19ff42",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vendorCode",
                                    "gameCode",
                                    "userCode",
                                    "language"
                                ],
                                "properties": {
                                    "vendorCode": {
                                        "type": "string",
                                        "example": "PGS"
                                    },
                                    "gameCode": {
                                        "type": "string",
                                        "example": "fortune-tiger"
                                    },
                                    "userCode": {
                                        "type": "string",
                                        "example": "user123"
                                    },
                                    "language": {
                                        "type": "string",
                                        "example": "en"
                                    },
                                    "lobbyUrl": {
                                        "type": "string",
                                        "example": "https://example.com/lobby",
                                        "nullable": true
                                    },
                                    "theme": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Launch URL retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "https://game.example.com/launch?token=..."
                                        },
                                        "errorCode": {
                                            "type": "integer",
                                            "example": 0
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/game/user/set-rtp": {
            "post": {
                "tags": [
                    "RTP"
                ],
                "summary": "Set user RTP for a vendor",
                "operationId": "704c8adf0eb11687390e1fcdb9773ef1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vendorCode",
                                    "userCode",
                                    "rtp"
                                ],
                                "properties": {
                                    "vendorCode": {
                                        "type": "string",
                                        "example": "PGS"
                                    },
                                    "userCode": {
                                        "type": "string",
                                        "example": "user123"
                                    },
                                    "rtp": {
                                        "type": "integer",
                                        "example": 85,
                                        "maximum": 99,
                                        "minimum": 30
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "RTP set successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "errorCode": {
                                            "type": "integer",
                                            "example": 0
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/game/user/get-rtp": {
            "post": {
                "tags": [
                    "RTP"
                ],
                "summary": "Get user RTP for a vendor",
                "operationId": "c31e5ce96e3edb1907ae89808f45ba91",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vendorCode",
                                    "userCode"
                                ],
                                "properties": {
                                    "vendorCode": {
                                        "type": "string",
                                        "example": "PGS"
                                    },
                                    "userCode": {
                                        "type": "string",
                                        "example": "user123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User RTP retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "integer",
                                            "example": 85
                                        },
                                        "errorCode": {
                                            "type": "integer",
                                            "example": 0
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/game/users/reset-rtp": {
            "post": {
                "tags": [
                    "RTP"
                ],
                "summary": "Reset RTP for all users of a vendor",
                "operationId": "c9d176ed6cbad9d9ad4cb0e841da0464",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vendorCode",
                                    "rtp"
                                ],
                                "properties": {
                                    "vendorCode": {
                                        "type": "string",
                                        "example": "PGS"
                                    },
                                    "rtp": {
                                        "type": "integer",
                                        "example": 85,
                                        "maximum": 99,
                                        "minimum": 30
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "RTP reset for all users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "integer",
                                            "example": 85
                                        },
                                        "errorCode": {
                                            "type": "integer",
                                            "example": 0
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/game/users/batch-rtp": {
            "post": {
                "tags": [
                    "RTP"
                ],
                "summary": "Batch set RTP for multiple users",
                "operationId": "d52b3818f07e8a6b0cfee7c431986e05",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vendorCode",
                                    "data"
                                ],
                                "properties": {
                                    "vendorCode": {
                                        "type": "string",
                                        "example": "PGS"
                                    },
                                    "data": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "userCode": {
                                                    "type": "string",
                                                    "example": "user123"
                                                },
                                                "rtp": {
                                                    "type": "integer",
                                                    "example": 85
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Batch RTP set successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "errorCode": {
                                            "type": "integer",
                                            "example": 0
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token obtained from /api/v2/auth/createtoken",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "basicAuth": {
                "type": "http",
                "description": "Basic Auth for Seamless Wallet callbacks",
                "scheme": "basic"
            }
        }
    },
...
}